finish parse, start vp track
finished parsing function, and started vicotry pt tracking

+++ main.cpp
@@ -44,7 +44,7 @@ void print_summary() {
         cout << setw(20) << cut_name << endl;
         for (auto const& c : p.second.cards_in_deck)
         {
-            cout << '\t' << setw(20) << c.first << " " << c.second << endl;
+            cout << '\t' << setw(20) << (int)c.first << " " << c.second << endl;
         }
     }
